GNSS: a wide hAcc threshold wrapped to 70 cm and rejected every fix
hacc_filter_threshold is an unsigned int of metres; the comparison against
nav.pvt.hAcc needs millimetres, so the code multiplied by 1000 in 32 bits.
Above 4294967 m that wraps. An operator setting 4294968 -- the natural way
to say "no filtering, keep everything" -- got 704 mm effective instead, and
the receiver silently discarded every real fix. Nothing in the log said so;
the beacon just stopped reporting positions.

Widen the comparison to 64 bits, and bound the three thresholds that feed
it (GNP21, LBP10, ZOP16) to 100000 m, which is four times the worst hAcc a
M10Q will ever report and leaves the wrap unreachable. They were declared
0..0, i.e. unbounded, so the DTE port accepted the overflowing value in the
first place.

Third instance of this `* 1000` pattern, after the rate limiter.

+++ ports/nrf52840/core/hardware/m10qasync/m10qasync.cpp
@@ -1007,7 +1007,15 @@ void M10QAsyncReceiver::react(const UBXCommsEventNavReport &n) {
 				    break;
 			    }
 
-			    if (m_nav_settings.hacc_filter_en && (m_nav_settings.hacc_filter_threshold * 1000) < nav.pvt.hAcc) {
+			    // 64-bit on purpose. hacc_filter_threshold is an unsigned int of
+			    // METRES and the comparison needs millimetres, so `* 1000`
+			    // overflowed above 4294967 m: a threshold of 4294968 wrapped to
+			    // 704 mm effective, and the receiver then rejected every real fix
+			    // while the operator believed he had disabled the filter. Nothing
+			    // was logged -- the beacon simply stopped reporting positions.
+			    // Same class as the rate-limiter `* 1000` fixed in 2026-08.
+			    if (m_nav_settings.hacc_filter_en &&
+			        ((uint64_t)m_nav_settings.hacc_filter_threshold * 1000u) < nav.pvt.hAcc) {
 				    // Fix exists but fails hAcc filter — store as degraded if best so far
 				    if (!m_has_degraded_pvt || nav.pvt.hAcc < m_degraded_pvt.hAcc) {
 					    m_degraded_pvt = { .iTOW = nav.pvt.iTow,
